requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
charset-normalizer==3.3.2
crispy-bootstrap5==2024.10
Django==4.2.16
django-crispy-forms==2.3
django-ses==4.2.0
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==1.26.20
